[MIG] sale_order_lot_generator: Migration to 14.0
