+++ sale_order_lot_generator/__manifest__.py
@@ -1,15 +1,14 @@
-#   @author Valentin CHEMIERE
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 {
     "name": "Sale Order Lot Generator",
-    "version": "12.0.1.0.1",
+    "version": "14.0.1.0.0",
     "author": "Akretion, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/sale-workflow",
     "license": "AGPL-3",
     "category": "Sale",
     "depends": ["sale_order_lot_selection"],
     "maintainers": ["florian-dacosta", "mourad-ehm", "bealdav"],
-    "data": ["views/product_view.xml"],
+    "data": ["views/product_template.xml"],
     "installable": True,
 }
