+++ sale_order_lot_generator/models/__init__.py
@@ -1,2 +1,3 @@
-from . import product
-from . import sale
+from . import product_template
+from . import sale_order
+from . import sale_order_line
