+++ sale_order_lot_generator/models/product_template.py
@@ -1,4 +1,3 @@
-#  @author Adrien CHAUSSENDE <adrien.chaussende@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from odoo import fields, models
