+++ sale_order_lot_generator/views/product_template.xml
@@ -5,9 +5,12 @@
         <field name="model">product.template</field>
         <field name="inherit_id" ref="stock.view_template_property_form" />
         <field name="arch" type="xml">
-            <div name="weight" position="after">
-                <field name="auto_generate_prodlot" />
-            </div>
+            <field name="tracking" position="after">
+                <field
+                    name="auto_generate_prodlot"
+                    attrs="{'invisible': [('tracking', '=', 'none')]}"
+                />
+            </field>
         </field>
     </record>
